Start making the app core modular
