+++ experiments/soup-app-demo/Cargo.toml
@@ -4,54 +4,39 @@ name = "soup-app-demo"
 version = "0.1.0"
 license = "MIT OR Apache-2.0"
 
-[features]
-default = ["nightly"]
-nightly = [
-    "embassy-executor/nightly",
-    "embassy-nrf/nightly",
-]
-
 [dependencies.embassy-futures]
 version = "0.1.0"
-# path = "../../embassy-futures"
 git = "https://github.com/embassy-rs/embassy"
-rev = "18fe398673f55b07159d01a230910bb9689c1525"
+rev = "7a841b58d127cc6d22c8895197d3f4d4c0974ad7"
 
 [dependencies.embassy-executor]
 version = "0.1.0"
-# path = "../../embassy-executor"
 git = "https://github.com/embassy-rs/embassy"
-rev = "18fe398673f55b07159d01a230910bb9689c1525"
-features = ["integrated-timers"]
+rev = "7a841b58d127cc6d22c8895197d3f4d4c0974ad7"
+features = ["nightly", "integrated-timers"]
 
 [dependencies.embassy-nrf]
 version = "0.1.0"
-# path = "../../embassy-nrf"
 git = "https://github.com/embassy-rs/embassy"
-rev = "18fe398673f55b07159d01a230910bb9689c1525"
-features = ["nrf52840", "time-driver-rtc1", "gpiote", "unstable-pac", "time"]
+rev = "7a841b58d127cc6d22c8895197d3f4d4c0974ad7"
+features = ["nightly", "nrf52840", "time-driver-rtc1", "gpiote", "unstable-pac", "time"]
 
 [dependencies.embassy-usb]
 version = "0.1.0"
-# path = "../../embassy-usb"
 git = "https://github.com/embassy-rs/embassy"
-rev = "18fe398673f55b07159d01a230910bb9689c1525"
+rev = "7a841b58d127cc6d22c8895197d3f4d4c0974ad7"
 features = ["msos-descriptor",]
-# optional = true
 
 [dependencies.embassy-time]
 version = "0.1.0"
 git = "https://github.com/embassy-rs/embassy"
-rev = "18fe398673f55b07159d01a230910bb9689c1525"
+rev = "7a841b58d127cc6d22c8895197d3f4d4c0974ad7"
 features = []
 
 [dependencies.embassy-sync]
 version = "0.1.0"
-# NOTE: using my version to include https://github.com/embassy-rs/embassy/pull/1296
-# git = "https://github.com/embassy-rs/embassy"
-# rev = "18fe398673f55b07159d01a230910bb9689c1525"
-git = "https://github.com/jamesmunns/embassy"
-rev = "4b5581be96f549cbab2f464f4eb5167ee4ab5ed1"
+git = "https://github.com/embassy-rs/embassy"
+rev = "7a841b58d127cc6d22c8895197d3f4d4c0974ad7"
 features = []
 
 [dependencies.soup-icd]
@@ -60,6 +45,7 @@ path = "../../shared/soup-icd"
 [dependencies]
 cortex-m = { version = "0.7.6", features = ["critical-section-single-core"] }
 cortex-m-rt = "0.7.0"
+heapless = "0.7.16"
 panic-reset = "0.1.1"
 postcard = "1.0"
